Initial draft of primitive type declaration embedding into Dhall
The core idea is to represent primitive types as a kind, and as a
function or just a simple sum type with one constructor.  When type
definition embedding into Dhall is done, value representation of a
primitive will be used instead of its name.  This will allow us to type
check the type definitions, and make sure that they are well formed.